Implements: https://github.com/aws/aws-rfdk/issues/366

### Testing

Started up a basic render farm using the RFDK examples, but modified the example to set the `fsc` mount option on the Amazon EFS and enable `cachefilesd` on the RenderQueue.

Used systems manager connection to remote in to the RenderQueue's ECS container host, and then:
1. Verified that cachefilesd is installed and running.
2. Verified that the `/mnt/repo` filesystem was mounted using the EFS helper which in-turn uses the NFSv4 mount driver, and that the `fsc` option was properly set on the NFSv4 mount point.
```bash
[root@ip-10-0-126-223 ~]# mount | grep nfs
sunrpc on /var/lib/nfs/rpc_pipefs type rpc_pipefs (rw,relatime)
127.0.0.1:/ on /mnt/repo type nfs4 (rw,relatime,vers=4.1,rsize=1048576,wsize=1048576,namlen=255,hard,noresvport,proto=tcp,port=20312,timeo=600,retrans=2,sec=sys,clientaddr=127.0.0.1,fsc,local_lock=none,addr=127.0.0.1,_netdev)
```
3. Ran a series of full copies of the `/mnt/repo` filesystem (approx 1.3 GB of data) to the local device both with and without cachefilesd running. The image below shows the throughput metrics on the Amazon EFS filesystem through these tests; as expected, the read throughput with cachefilesd is much lower than a full copy once the local cache has been populated.

![RepoCopy](https://user-images.githubusercontent.com/53624638/112866441-6bdfaf80-907f-11eb-9ca3-41c01f5c4ea9.png)

### Performance Testing

Test setup:
1. RFDK Basic all-in application with an added bastion host set up as a Workstation (installed Deadline & DCV). So... single RCS (c5.large), EFS burst-mode filesystem with ~1GB of data (encrypted), DocDB (db.r5.large), Linux-based Worker ASG (t3.medium), and TLS enabled throughout

![summary](https://user-images.githubusercontent.com/53624638/113048881-9fe2cf80-9168-11eb-9fee-a24831c863fd.png)

So, cachefilesd shaves about 30% of metered throughput when starting 200 workers (9.73MB/s -> 6.9MB/s) and 26% of metered throughput when rendering 1k 'sleep 10' jobs (1.9MB/s -> 1.6MB/s). With the standard caveats, of course, about this being a single data point so there will be sizable error bars on those.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*