@RandomInsano 2021-03-29
Looks great! My only concern is with the namespace for this option and underlying function names being generic as it has the following two effects:
1. Those aware of cachefilesd have to check the docs to know what the underlying tech is
2. If we find an *even better* caching option or choose to change the Repo's backing protocol (say to SMB) we'll have problems.

Maybe we could swap "enable_local_file_caching" for "enable_fsc_caching"?