@ddneilson 2021-03-31
> 
> 
> Looks great! My only concern is with the namespace for this option and underlying function names being generic as it has the following two effects:
> 
>     1. Those aware of cachefilesd have to check the docs to know what the underlying tech is
> 
>     2. If we find an _even better_ caching option or choose to change the Repo's backing protocol (say to SMB) we'll have problems.
> 
> 
> Maybe we could swap "enable_local_file_caching" for "enable_fsc_caching", etc?

Thanks for the feedback, Edwin. Variable/property naming is hard. ;-) On one hand... generic for those that don't know the details, or don't want to know/care... On the other hand... it is nice to know what is going on under the hood. I'm trying to strike that balance by going generic with the property name and providing details in the docstring (which shows up under quick-info-type IDE functionality). My thinking is also that if there's a different caching tech for a different filesystem, then we'd want to support that under the same option but detect the filesystem type if we can; it's all the same high-level functionality, regardless of how it's accomplished.

You're thinking that I'm missing the mark with it?